rename: DataSource->DataContainer, add CircuitContainer
the method is called addSource, but the class is called
Plasma::DataContainer... strange ;P but well, we are now in sync again.

also add an implementation of a DataContainer that gets information from
a circuit and stores it in a Plasma::DataContainer way. Still needs to
be implemented, does nothing for now.