add new ComponentContainer class
This is a Plasma::DataContainer that will provide access to semantic
information included within SVG files describing KTechLab-components.

This includes a dependency on Plasma for src/interfaces, which isn't
optimal. I think, in the future there will be a new Plugin to handle
components in a generic manner (i.e. flow-chart components as well as
electronic components) defined in SVG-files.