implement more ktlproject functionality
This also fixes a bug with item names in a project, which aren't unique.
This should now be handled correctly.

I'm not sure, where to actually rename the files in the filesystem, so I
haven't implemented that, yet. It seems, this isn't done by
KDevPlatform, so this is still needed. The project file is updated just
fine.

There is a problem with removing folders still containing files. They
won't be removed from the filesystem, because they are not empty. Need
to find out, how this is handled and what we can do about that.