provide default implementation in IConRouter
Implement IConRouter::mapRoute(qreal,qreal,qreal,qreal) directly in the
abstract class. This way the implementation only must override one
method instead of both. Since we expect the default implementation to
work on QPointF to represent a point, the method working with single
coordinates is implemented in a default way.