make components svg files, instead of svgz
this has some advantages:
- changes can be tracked from now
- more easy to parse the svg DOM to extract information

disadvantages:
- larger file-size

Since these files are somewhat "part" of the program itself it is good
to be able to track changes in detail. Since we want to extract
information from the svg DOM (like position of pins and some css
information), it's okay to save these svg files un-compressed.