move all flowcode related files into plugins dir
These files will be ported later, so they shouldn't be deleted, yet.
However, they "pollute" the src/ directory, so I moved them into a
dedicated directory in src/plugins. Later a flowcode plugin will be
created to bring this feature back into KTechLab and these classes can
be used as a base to start working on.