make dataUpdated and updateData consistent
This updates the signatures of methods that are used to exchange
information about data updates between model and view to make them more
consistent.

There is now a public slot updateData(name,data) which is virtual and
a signal dataUpdated(name,data) which will be emitted, when some data is
changed.