implement visualizedData for automatic_router
This helps debugging the internal data of the plugin and already
revealed some issues.

This somehow doesn't work as expected. This might
be a problem in Qt or in the way I use it. Sometimes the foreground
isn't painted, so the painted items overlap this data. As a work-around
I added some code that calls update() in the IDocumentScene, when the
right mouse-button is pressed. This will repaint the scene and the
foreground will be drawn correctly.