add alignToGrid-logic and use it when moving items
The scene will handle all moving of items, and it will optionally align
all items to a grid. by default, this will be handled by the routing
plugin, but this can be changed by sub-classing IDocumentScene. Also the
AutomaticRouter plugin aligns the objects to the 8x8-grid used
internally by the plugin.