cleanup automatic_router plugin
this is a pretty large commit that fixes some issues in automatic_router
plugin and makes it basically work, if the internal scores are all
up-to-date. routing will then behave very similar to the kde3 version.

there have been some misunderstandings with the coordinate-systems, so
these bugs are also fixed, now. some changes to the circuit plugin were
necessary to make this work. the components had to be moved by 4px as
well as had the connectors.

all in all the code in automatic_router is now cleaned up a bit and
works, kind of. there is still some work to be done update the internal
scores to represent the actual scene.