unify pin id definition
the ids must be compatible in simulator and visualisation. the pin ids
should be defined either in the component or the element but not both.
the component also knows about the pins, because of ids found in the
svg-files. this information can be used by the elements as shown below.