refactor CircuitModel
first steps towards the new circuit model. it will be a tree like model,
because the documents are stored in XML. for now, this will make
visualisation of old circuits crash, but this will be fixed, soon.
it is still able to load circuit files and a first test for that is now
implemented.