fix the releasing of sliders on component when mouse button is released
before, after starting to drag the slider of a component with has one, the slider would also move when the mouse button is released. this is fixed now.
internally, the mouse release event has been incorrectly generated and it signaled that the mouse button is released but it is also being pressed.