process chain: keep temporary files live longer
A QTemporaryFile on the stack is created just to get a temporary file
name; by default QTemporaryFile autodeletes the temporary file on
removal, so a step after an indirect one fails because the result of the
previous step does not exist anymore.

While setting setAutoRemove(false) would solve the problem, it has the
side effect of leaving the temporary file around. As a better solution,
create the QTemporaryFile on the heap, parenting it to the processor
class: as the processor is removed only when the process chain ends,
the temporary file will be around during the whole chain processing,
being removed at the end.