Rewrite SerialPort to use QSerialPort
Switch the SerialPort implementation to QSerialPort instead of the
custom Unix-only implementation, so it can work on many more platforms
than before. Drop the probe() method, used only internally by ports(),
and thus no more needed. Since QSerialPort cannot set the DSR pin, the
only Unix-only bit left is for that.

Adapt the type for the baud rate to qint32, and specify a value
directly in SerialPortComponent, so the termios usage is removed.

Fixes #5