DrawingPushButton -> QPushButton
QPushButton can show a pixmap just fine, so there is no need for a
special QPushButton subclass. Let's set the icon size we want, and
the pixmap-as-icon will be shown automatically.

Also get rid of the manual calls to update(), as changes to the
properties of a widget already trigger its update (since Qt 4 already,
IIRC).