Remove almost all the 'const char *name' params
This was needed up to Qt 3, because the only way to set an object name
was to pass it to the QObject constructor. Qt 4 and on is way more
flexible about this, allowing to change an object name anytime.

Hence, remove almost all the 'const char *name' parameters in
constructors, and functions that end up invoking constructors; instead,
the object name is set mostly by the caller.

Exceptions to this mass change:
- ktlqt3support (which is mostly Qt 3 bits)
- SettingsDlg, as it is needed for KConfigDialog