SerialPort: unroll setPinState() & pinState()
Half of the values of the current Pin enum can be used only with
setPinState() and the other half only with pinState(), making it
somehow puzzling as single enum.

Instead, unroll setPinState() and pinState() to single functions to
set bits and get the values of line signals, removing the Pin enum.
This simplifies the code a bit, also making the SerialPort API clearer
to use.