Fix crash when opening a document using KRecentFilesAction
When opening a document from the slot triggered by KRecentFilesAction,
it is a bad idea to call functions of the same KRecentFilesAction that
triggered it, leading to corruption/crash.

As solution, split the loading of files from KRecentFilesAction to a
separate slot, different than the general function used for loading
files: this way, that code path will not call back functions of
KRecentFilesAction. An helper function helps sharing the common code.

# Conflicts:
#	src/ktechlab.cpp
#	src/ktechlab.h